replaced TopTenComponent with HomePage component
