+++ .gitignore
@@ -4,6 +4,7 @@
 .*.swp
 .*.swo
 # See http://help.github.com/ignore-files/ for more about ignoring files.
+secrets.ts
 
 # compiled output
 /dist
