--- src/app/TopTenTable/TopTenTable.component.html
@@ -1,21 +1,0 @@
-<h2>Current Top 10</h2>
-<div class="selectContainer">
-  <h6>Sort By:</h6>
-  <p class="textCenter">Beer Type</p>
-  <select class="selectBoxBeerType" for="beerType">
-    <option *ngFor="let beer of beerTypes" value="{{ beer.type }}">
-      {{ beer.type }}
-    </option>
-  </select>
-</div>
-<table class="myTable">
-  <tr>
-    <th class="tableHeaderName" mat-sort-header="name">Name</th>
-    <th mat-sort-header="rating">Rating</th>
-  </tr>
-
-  <tr *ngFor="let beer of REVIEWS">
-    <td class="tableItemName">{{ beer.beerName }}</td>
-    <td>{{ beer.rating }}/10</td>
-  </tr>
-</table>
