+++ src/app/app.component.html
@@ -14,6 +14,7 @@
   <div class="collapse navbar-collapse" id="navbarNav">
     <ul class="navbar-nav">
       <li class="nav-item">
+        <!-- TODO: remove homepage -->
         <a class="nav-link" routerLink="/homepage"
           >Home<span class="sr-only">(current)</span></a
         >
