+++ src/app/homepage/homepage.component.css
@@ -1,0 +1,37 @@
+
+h2 {
+  text-align: center;
+  padding-bottom: 20px;
+}
+
+.myTable {
+  padding-top: 20px;
+  align-items: center;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.tableHeaderName {
+  float: left;
+}
+
+.tableItemName {
+  padding-right: 40px;
+}
+
+h6 {
+  text-align: center;
+}
+.selectBoxBeerType {
+  margin-right: auto;
+  margin-left: auto;
+  display: block;
+}
+
+.selectContainer {
+  padding-bottom: 25px;
+}
+
+.textCenter {
+  text-align: center;
+}
