+++ src/app/shared/footer/footer.component.html
@@ -3,6 +3,7 @@
 <footer class="bg-dark text-center text-lg-start">
   <div>
     <nav>
+      //TODO: Home
       <a class="linkSeperator" routerLink="/homepage">Home</a>
       <a class="linkSeperator" routerLink="/review-page">Reviews</a>
       <a class="linkSeperator" routerLink="/merch-page">Merch</a>
